+++ libgeo/opf.c
@@ -1,5 +1,5 @@
 /* See rn_classif_opf.h. */
-/* Last edited on 2024-11-20 15:44:07 by stolfi */
+/* Last edited on 2024-11-22 03:51:28 by stolfi */
 
 #include <stdio.h>
 #include <stdlib.h>
@@ -25,7 +25,7 @@ void opf_build_complete
         
     /* Allocate a queue and put all vertices in it with null predecessors: */
     uint32_t *Q = talloc(n, uint32_t); /* Sample index queue. */
-    for (int32_t u = 0; u < n; u++) { Q[u] = u; P[u] = u; }
+    for (int32_t u = 0; u < n; u++) { Q[u] = (uint32_t)u; P[u] = (uint32_t)u; }
 
     auto void resortQ(uint32_t k0, uint32_t k1);
       /* Rearranges {Q[k0..k1-1]} so that {C[Q[k0]]} is minimum. */
@@ -47,7 +47,7 @@ void opf_build_complete
         R[u] = (P[u] == u ? u : R[P[u]]);
         m++;
         /* Update the cost of all remaining items in {Q}: */
-        for (int32_t k = m; k < n; k++)
+        for (uint32_t k = m; k < n; k++)
           { uint32_t v = Q[k]; 
             if (verbose) { fprintf(stderr, "    checking  %5d cost = %8.6f", v, C[v]); }
             double Ca = acost(v, u); /* Cost of arc {(v,u)}. */
@@ -70,7 +70,7 @@ void opf_build_complete
         uint32_t t = k0;
         uint32_t u = Q[t];
         double Cup = C[u];
-        for (int32_t k = k0 + 1; k < k1; k++) 
+        for (uint32_t k = k0 + 1; k < k1; k++) 
           { if (C[Q[k]] < Cup) { t = k; u = Q[t]; Cup = C[u]; } }
         if (t != k0) { Q[t] = Q[k0]; Q[k0] = u; } 
       }
